requirements.txt
opencv-python
requests
gpiozero
RPLCD
RPi.GPIO